refactor: add file with utilites and folder with types , clean unused methods in index.ts, renders.ts, store.ts
