+++ src/types/storeTypes.ts
@@ -1,0 +1,7 @@
+import { CarObject } from '../types/apiTypes'
+
+export type State = {
+  carsPage: number,
+  cars: CarObject[],
+  carsCount: number,
+}
