+++ src/utils.ts
@@ -1,0 +1,4 @@
+export const randomColor = () => {
+  return "#" + Math.floor(Math.random()*16777215).toString(16).padStart(6, '0').toUpperCase();
+}
+
